+++ noodles-vcf/examples/vcf_count.rs
@@ -2,23 +2,19 @@
 //!
 //! The result matches the output of `bcftools view --no-header <src> | wc -l`.
 
-use std::{
-    env,
-    fs::File,
-    io::{self, BufReader},
-};
+use std::{env, fs::File, io::BufReader};
 
 use noodles_vcf as vcf;
 
-fn main() -> io::Result<()> {
+fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = File::open(src).map(BufReader::new).map(vcf::Reader::new)?;
-    reader.read_header()?;
+    let header = reader.read_header()?.parse()?;
 
     let mut n = 0;
 
-    for result in reader.records() {
+    for result in reader.records(&header) {
         let _ = result?;
         n += 1;
     }