+++ noodles-vcf/src/reader/records.rs
@@ -1,30 +1,31 @@
 use std::io::{self, BufRead};
 
-use crate::Record;
-
 use super::Reader;
+use crate::{Header, Record};
 
 /// An iterator over records of a VCF reader.
 ///
 /// This is created by calling [`Reader::records`].
-pub struct Records<'a, R> {
-    inner: &'a mut Reader<R>,
+pub struct Records<'r, 'h, R> {
+    inner: &'r mut Reader<R>,
+    header: &'h Header,
     line_buf: String,
 }
 
-impl<'a, R> Records<'a, R>
+impl<'r, 'h, R> Records<'r, 'h, R>
 where
     R: BufRead,
 {
-    pub(crate) fn new(inner: &'a mut Reader<R>) -> Self {
+    pub(crate) fn new(inner: &'r mut Reader<R>, header: &'h Header) -> Self {
         Self {
             inner,
+            header,
             line_buf: String::new(),
         }
     }
 }
 
-impl<'a, R> Iterator for Records<'a, R>
+impl<'r, 'h, R> Iterator for Records<'r, 'h, R>
 where
     R: BufRead,
 {
@@ -36,8 +37,7 @@ where
         match self.inner.read_record(&mut self.line_buf) {
             Ok(0) => None,
             Ok(_) => Some(
-                self.line_buf
-                    .parse()
+                Record::try_from_str(&self.line_buf, self.header)
                     .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)),
             ),
             Err(e) => Some(Err(e)),