+++ noodles-vcf/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Changed
 
+  * vcf/reader: `Reader::records` now takes a `Header` to use for record
+    parsing.
+
   * vcf/record/format: Prefer key from header.
 
   * vcf/record/info/field: Prefer key from header.